llmops-api/requirements.txt
Flask==3.1.2
flask_wtf==1.2.2
injector==0.22.0
openai==1.109.1
python-dotenv==1.1.1
WTForms==3.2.1
